Develop getPlayerChoice function, make correction
Correct the choice variable in getComputerChoice() function. Change variable
in line 5 to computerChoice.
Develop getPlayerChoice() function. If user cancels the prompt, alert
with 'Game is over!'. If user inputs an invalid choice, alert it. If user
input either 'Rock', 'Paper' or 'Scissors', case insensitive,
convert the input to uppercase and compare it with 'ROCK', 'PAPER',
'SCISSORS' to make sure the input is valid, then return the input.