Create playRound function
Create playRound function which takes 2 parameters: user's selection
and computer's selection. Input 2 conditions to the function so far.
If user's choice is the same as the computer's, return 'It's a draw!'.
If user's choice is rock and the computer's: scissors, return 'You
win! Rock beats Scissors.'.
Run function only if user inputs a valid value.