Create game function, add variables in playRound
Create 2 variables: playerScore and computerScore and initialise to 0.
Increment the variables in playRound() function. If player wins, increment
playerScore by 1. If computer wins, increment computerScore by 1.
Create game() function. Add for loop which will run for 5 rounds and calls
the playRound function. If player cancels the game, break the loop and end
the game. Else compare playerScore with computerScore in the end and
announce winner or loser or if the game is a draw.
If player inputs nothing and click ok, the game is ending instead of
alerting 'Invalid input'. To check.
First round is not being played. To check.