Edit the game and playRound functions
Change the playRound() functions results. Change from return to console.log.
Move playerSelection and computerSelection variables in game()
function, in the loop. Add condition to play the round if the player
inputs a valid value. It does not work as it should yet.
Make sure the game stops when the player clicks on cancel, and prompts
the player again for an input if the player inputs an empty value.
For the moment both of these scenarios end up in cancelling the game.