Add event listener to buttons and call playRound
Comment out the game() and the capiraliseWord() functions. Add an
event listener for each of the three buttons created in the index.html.
Call the playRound() function each time one of the buttons is clicked.
The playRound() will take the button's value as the user's input and
the computer's choice between rock, paper and scissors. Take the
user's input via the buttons and not the prompt.