Remove unneeded functions and update playRound()
Remove the function prompting the user to enter their input. Remove
capitalise function which capitalised the user's input. Remove the
game() function.
Edit the playRound() function. Create new variable final and assign it
to the final div in index.html. Add condition to announce the winner of
the game once the score of either player reaches 5.