Change and add condition spot within playRound()
Change condition, which checks whether playerScore or computerScore
is greater than 4, to the last part of the playRound() function.
Add condition in the first part of playRound(): only execute the function
if the text content of the final div is an empty string. The game will
end once a winner is announced.
Move around some variables.